requirements.txt
pexpect==4.9.0
pyotp==2.9.0
python-dotenv==1.0.1
pyyaml==6.0.1
